Refactored kernel state tests.
